Extract Formatting into separate file.
Extract Formatting struct, Terminal struct, and Color.Wrap extensions into `Formatting.swift`.
